Make a metric only go to one copier at a time
Change the communication pattern so that the batcher will only
send it's data to one copier at a time until the copy is done.

This avoids the possibility that two copiers will write data to the
same table at the same time, avoiding contention.

This may be bad if we need parallelism writing to a single metric.